Cleaned code and improved redux structure
