+++ dashboard/src/components/Wrapper.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-const Wrapper = function Wrapper({children}) {
+const Wrapper = function Wrapper({ children }) {
 
   return(
     <div className="flex justify-center items-center h-100">
