+++ dashboard/src/components/chat/Chat.jsx
@@ -4,7 +4,9 @@ const Chat = function Chat({ chatMessages }) {
 
   return(
     <div>
-
+      {chatMessages.map((msg) =>
+        <p>{msg}</p>
+      )}
     </div>
   )
 }
