--- dashboard/src/components/chat/redux/store.jsx
@@ -1,4 +1,0 @@
-import { createStore } from 'redux'
-import { messageHandler } from './reducers'
-
-export const store = createStore(messageHandler, { message: 'test', alt: "asd" })
