+++ dashboard/src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import MainLayout from './layouts/MainLayout.jsx';
 import registerServiceWorker from './registerServiceWorker';
 import { Provider } from 'react-redux';
-import { store } from './store';
+import { store } from './mainRedux/store';
 import './index.css';
 
 ReactDOM.render(
