+++ dashboard/src/mainRedux/store.js
@@ -1,0 +1,4 @@
+import { createStore } from 'redux';
+import Reducers from './mainReducers';
+
+export const store = createStore(Reducers)
