--- dashboard/src/store.js
@@ -1,4 +1,0 @@
-import { createStore } from 'redux'
-import Reducers from './reducers/index.js'
-
-export const store = createStore(Reducers)
